infra: on-cluster scoring, float32 pipeline, A2 sweep + experiment roadmap
Analysis moves to the cluster so only kilobytes travel home (the N=12500
matrices are ~1 GB each and naive scoring OOMs a laptop):
- scripts/analyze_run.py — loads one matrix at a time, scores all off-diagonal
  edges when small enough else a 20M-edge random sample, writes metrics.csv.
  Verified on the N=12500 probe: reproduces AUC 0.620 without OOM.

float32 end-to-end:
- streaming.py chunk pipeline (spikes/noise/AR) in float32, standard_normal*sigma
  instead of normal(0,sigma) — less memory traffic, faster at large N.
- estimates saved as float32: 1.25 GB -> 0.6 GB per matrix.
- validated: OLS AUC 0.6203 vs 0.6210 in float64 (negligible), numpy and torch
  paths identical.

slurm/run_n12500ai_gpu.slurm is now experiment A2: one 5M-ms sim per seed scored
at 1M/2M/5M checkpoints, then on-cluster analysis; 8 h wall.

docs/experiments/experiment_roadmap.md records the full prioritized experiment
plan (data length & scaling, bias-variance, regime, signal model, partial
observation) so nothing is lost.

+++ scripts/wrapup_run_stream.py
@@ -114,9 +114,11 @@ def main():
             mats = solve_selected(Cxx, Cyx, cfg, args.methods)
             outdir = Path(base) / "results" / f"{cfg['name']}_T{int(T)//1000}k" / f"seed{seed}"
             outdir.mkdir(parents=True, exist_ok=True)
-            np.save(outdir / "adj_true.npy", adj_true)
+            # float32 on disk: halves file size (1.25 GB -> 0.6 GB at N=12500)
+            # with no meaningful precision loss for connectivity analysis.
+            np.save(outdir / "adj_true.npy", adj_true.astype(np.float32))
             for name, A in mats.items():
-                np.save(outdir / f"A_{name}.npy", A)
+                np.save(outdir / f"A_{name}.npy", A.astype(np.float32))
             k0 = next(iter(mats))
             print(f"[seed {seed}] T={T:.0f}ms done -> {outdir}  "
                   f"|{k0}|max={np.abs(mats[k0]).max():.2e}", flush=True)
